Enhance packaging scripts to handle public releases on tags

+++ scripts/create_all_packages.sh
@@ -3,6 +3,18 @@ BASEDIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"/..
 echo "BASEDIR is $BASEDIR"
 cd $BASEDIR
 
+# Reading arguments
+while getopts tdv: option; do
+    case "${option}" in
+        t) TAG="-t";; # Indicates that we're releasing a public version from a tag
+        d) DRYRUN="-d";; # Indicates a dryrun test, that does not modify anything on the network
+        v) VERSION=${OPTARG};; # Gives the version string to use (else it will use the commit id)
+    esac
+done
+
+MAJOR_NUMERIC_VERSION="2.1"
+VERSION_TO_REPLACE="2\.1-WIP"
+
 # Set the secret environment variables if available
 # The file set_secret_environment_variables.sh should not be commited for security reasons
 # It is only useful to run the scripts locally.
@@ -11,15 +23,15 @@ if [ -r "$BASEDIR/scripts/set_secret_environment_variables.sh" ]; then
   . "$BASEDIR/scripts/set_secret_environment_variables.sh"
 fi
 
-# Use the first argument as a version number, else use the Travis tag, else use the commit id
-VERSION_TO_REPLACE="2\.0-WIP"
-VERSION_PREFIX="2.0"
-if [ ! "$1zz" == "zz" ]; then
-    VERSION=$1
+# Use the passed version number, else use the commit id
+if [ ! "${VERSION}zz" == "zz" ]; then
     echo "Packaging version $VERSION because it has been passed as an argument"
+    if [ ! "${TAG}zz" == "zz" ]; then
+        echo "This version is a tag : we're releasing a public version"
+    fi
 else
     COMMIT_ID=$(git rev-parse --short HEAD)
-    VERSION="${VERSION_PREFIX}commit-${COMMIT_ID}"
+    VERSION="${MAJOR_NUMERIC_VERSION}commit-${COMMIT_ID}"
     echo "Packaging version $VERSION"
 fi
 
@@ -34,7 +46,7 @@ regexpNumericVersion='^[0-9\.]+$'
 if [[ $VERSION =~ $regexpNumericVersion ]] ; then
    sed -i -e "s/$VERSION_TO_REPLACE/$VERSION/" tmp/manifest.json
 else
-   sed -i -e "s/$VERSION_TO_REPLACE/$VERSION_PREFIX/" tmp/manifest.json
+   sed -i -e "s/$VERSION_TO_REPLACE/$MAJOR_NUMERIC_VERSION/" tmp/manifest.json
 fi
 sed -i -e "s/$VERSION_TO_REPLACE/$VERSION/" tmp/manifest.webapp
 sed -i -e "s/$VERSION_TO_REPLACE/$VERSION/" tmp/www/index.html
@@ -42,16 +54,20 @@ sed -i -e "s/$VERSION_TO_REPLACE/$VERSION/" tmp/www/index.html
 mkdir -p build
 rm -rf build/*
 # Package for Chromium/Chrome
-scripts/package_chrome_extension.sh $VERSION
+scripts/package_chrome_extension.sh $DRYRUN $TAG -v $VERSION
 # Package for Firefox and Firefox OS
 # We have to put the real version string inside the manifest.json (which Chrome might not have accepted)
 # So we take the original manifest again, and replace the version inside it again
 cp manifest.json tmp/
 sed -i -e "s/$VERSION_TO_REPLACE/$VERSION/" tmp/manifest.json
-scripts/package_firefox_extension.sh $VERSION
-scripts/package_firefoxos_app.sh $VERSION
+scripts/package_firefox_extension.sh $DRYRUN $TAG -v $VERSION
+scripts/package_firefoxos_app.sh $DRYRUN $TAG -v $VERSION
 
-CURRENT_DATE=$(date +'%Y-%m-%d')
-# Upload the files on download.kiwix.org
-echo "Uploading the files on http://download.kiwix.org/nightly/$CURRENT_DATE/"
-scp -r -p -i scripts/travisci_builder_id_key build/* nightlybot@download.kiwix.org:/var/www/download.kiwix.org/nightly/$CURRENT_DATE
+if [ "${DRYRUN}zz" == "zz" ]; then
+    CURRENT_DATE=$(date +'%Y-%m-%d')
+    # Upload the files on download.kiwix.org
+    echo "Uploading the files on http://download.kiwix.org/nightly/$CURRENT_DATE/"
+    scp -r -p -i scripts/travisci_builder_id_key build/* nightlybot@download.kiwix.org:/var/www/download.kiwix.org/nightly/$CURRENT_DATE
+else
+    echo "Skipping uploading the files, because it's a dryrun test"
+fi
